+++ project/DB/airflow/dags/02_fsc_first.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timedelta
-# import requests
-# from bs4 import BeautifulSoup
+import requests
+from bs4 import BeautifulSoup
 import pandas as pd
 
 from airflow import DAG
@@ -8,10 +8,10 @@ from airflow.operators.python_operator import PythonOperator
 # from elasticsearch import Elasticsearch, helpers
 import re
 
+import os
 from package.fsc_crawling import crawling
 from package.fsc_extract import extract_main_content, extract_reason
 from package.vector_embedding import generate_embedding
-import os
 
 from opensearchpy import OpenSearch, helpers
 from dotenv import load_dotenv
@@ -32,16 +32,40 @@ client = OpenSearch(
 # Elasticsearch 인스턴스 생성 (Docker 내부에서 실행 중인 호스트에 연결)
 # es = Elasticsearch('http://host.docker.internal:9200')
 
-# Elasticsearch 인덱스 생성 (이미 존재하면 무시)
-try:
-    # es.indices.create(index='raw_data')
-    client.indices.create(index='raw_data')
-except Exception as e:
-    print(f"인덱스 생성 오류 또는 이미 존재: {e}")
+# Elasticsearch 인덱스 생성 또는 재설정 함수
+def create_or_update_index():
+    """Elasticsearch 인덱스를 생성 또는 갱신하여 '날짜' 필드를 date 타입으로 설정"""
+    # 인덱스가 이미 존재하면 삭제
+    # if es.indices.exists(index='raw_data'):
+    #     es.indices.delete(index='raw_data')
+    #     print("기존 인덱스 삭제 완료")
+    if client.indices.exists(index='Korean_Law_data'):
+        client.indices.delete(index='Korean_Law_data')
+        print("기존 인덱스 삭제 완료")
 
+    # 새로운 인덱스 생성 (날짜 필드를 date 타입으로 설정)
+    index_settings = {
+        "mappings": {
+            "properties": {
+                "제목": {"type": "text"},
+                "제목_vector" : {"type":"knn_vector", "dimension": 1536},
+                "날짜": {"type": "date"},
+                "URL": {"type": "text"},
+                "내용": {"type": "text"},
+                "내용_vector" : {"type":"knn_vector", "dimension": 1536},
+                "개정이유": {"type": "text"},
+                "개정이유_vector" : {"type":"knn_vector", "dimension": 1536},
+                "주요내용": {"type": "text"},
+                "주요내용_vector" : {"type":"knn_vector", "dimension": 1536}
+            }
+        }
+    }
+    # es.indices.create(index='raw_data', body=index_settings)
+    client.indices.create(index='Korean_Law_data', body=index_settings)
+    print("새로운 인덱스 생성 완료")
 
 def crawling_extract_df():
-    df = crawling(3)
+    df = crawling(10)
     
     if df is None or df.empty:
         print("크롤링된 데이터가 없습니다. 빈 DataFrame을 반환합니다.")
@@ -54,43 +78,16 @@ def crawling_extract_df():
 
     return df
 
-def get_existing_entries():
-    # 현재 날짜와 1년 전 날짜 계산
-    current_date = datetime.now()
-    one_year_ago = current_date - timedelta(days=365)
-    
-    # 1년 전부터 현재까지의 범위 쿼리 생성
-    query = {
-        "query": {
-            "range": {
-                "날짜": {
-                    "gte": one_year_ago.strftime("%Y-%m-%d"),
-                    "lte": current_date.strftime("%Y-%m-%d"),
-                    "format": "yyyy-MM-dd"
-                }
-            }
-        }
-    }
-    
-    # Elasticsearch에서 검색
-    # response = es.search(index="raw_data", body=query, size=10000)
-    response = client.search(index="raw_data", body=query, size=10000)
-    existing_entries = {(hit['_source']['제목'], hit['_source']['날짜'], hit['_source']['URL']) for hit in response['hits']['hits']}
-    
-    print(f"검색된 데이터 개수: {response['hits']['total']['value']}")
-    print(f"기존 데이터 수: {len(existing_entries)}")
-    return existing_entries
 
-def upload_new_data():
+def upload_data():
     df = crawling_extract_df()
-    existing_entries = get_existing_entries()
-
-     # 벡터 임베딩 생성
+    
+    # 벡터 임베딩 생성
     df['제목_vector'] = df['제목'].apply(generate_embedding)
     df['내용_vector'] = df['내용'].apply(generate_embedding)
     df['개정이유_vector'] = df['개정이유'].apply(generate_embedding)
     df['주요내용_vector'] = df['주요내용'].apply(generate_embedding)
-
+    
     actions = [
         {
             "_op_type": "index",
@@ -110,41 +107,47 @@ def upload_new_data():
             }
         }
         for _, row in df.iterrows()
-        if (row['제목'], row['날짜'], row['URL']) not in existing_entries
     ]
 
-    print(f"중복 제거 후 삽입할 데이터 수: {len(actions)}")
+    print(f"삽입할 데이터 수: {len(actions)}")
     
     if actions:
         # helpers.bulk(es, actions)
         helpers.bulk(client, actions)
-        print(f"{len(actions)}개의 새로운 데이터를 업로드했습니다.")
+        print(f"{len(actions)}개의 데이터를 업로드했습니다.")
     else:
-        print("새로운 데이터가 없습니다.")
+        print("업로드할 데이터가 없습니다.")
+
 
 # Airflow DAG 기본 설정
 default_args = {
-    'depends_on_past': False,  # 이전 작업에 의존하지 않음
-    'retries': 1,  # 실패 시 재시도 횟수
-    'retry_delay': timedelta(minutes=5),  # 재시도 간격
+    'depends_on_past': False,
+    'retries': 1,
+    'retry_delay': timedelta(minutes=5),
 }
 
 # Airflow DAG 정의
 with DAG(
-    'fsc_daily',  # DAG 이름
-    default_args=default_args,  # 기본 설정
-    description="입법예고 데이터를 중복 없이 Elasticsearch에 저장합니다.",  # 설명
-    schedule_interval='@daily',  # 하루마다 실행
-    start_date=datetime.now(),  # 시작 날짜
-    catchup=False,  # 과거 실행 건 무시
-    tags=['elasticsearch', 'crawl', 'finance']  # 태그
+    '02.Korean_Law_Index_data',
+    default_args=default_args,
+    description="입법예고/규정변경예고 데이터를 생성하고 적재합니다.",
+    schedule_interval='@monthly',
+    start_date=datetime.now(),
+    catchup=False,
+    tags=['elasticsearch', 'crawl', 'finance']
 ) as dag:
+
+    # Elasticsearch 인덱스 초기화 작업
+    initialize_index = PythonOperator(
+        task_id="initialize_elasticsearch_index",
+        python_callable=create_or_update_index,
+    )
 
     # 데이터 업로드 작업 정의
     upload_task = PythonOperator(
-        task_id="upload_new_data_to_elasticsearch",  # 작업 ID
-        python_callable=upload_new_data,  # 실행할 함수
+        task_id="upload_data_to_elasticsearch",
+        python_callable=upload_data,
     )
 
-    # DAG 실행
-    upload_task
+    # DAG 실행 순서 설정
+    initialize_index >> upload_task
